backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
bleach==6.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==40.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-allauth==0.53.1
django-cors-headers==3.13.0
django-simple-history==3.3.0
django-timezone-field==5.0
djangoql==0.17.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
docutils==0.19
drf-spectacular==0.26.1
ducktoolkit==1.0.3
dumb-init==1.2.5.post1
elementpath==4.0.1
environ==1.0
frozenlist==1.3.3
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
hashids==1.3.1
html5lib==1.1
idna==3.4
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
joblib==1.2.0
jsonpickle==1.4.2
jsonschema==4.17.3
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
npx==0.1.1
numpy==1.24.2
oauthlib==3.2.2
optional-django==0.3.0
oscrypto==1.3.0
packaging==23.0
Pillow==9.4.0
ply==3.11
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pypdf==3.5.0
pypng==0.20220715.0
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
react==4.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
sendgrid==6.9.7
sentry-sdk==1.16.0
six==1.16.0
slack-bolt==1.16.3
slack-sdk==3.20.1
smtpapi==0.4.12
sniffio==1.3.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tenacity==8.2.2
threadpoolctl==3.1.0
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
ua-parser==0.16.1
undecorated==0.3.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
xmlschema==2.2.1
yarl==1.8.2
